+++ CHANGELOG.md
@@ -1,7 +1,14 @@
 # Unreleased
 
+Breaking
+  * AZ-515: Option to use Azure CAF naming provider to name resources
+  * AZ-515: Require Terraform 0.13+
+
 Changed
   * AZ-572: Revamp examples and improve CI
+
+Updated:
+  * AZ-589: Update diagnostics settings to `v4.0.3`
 
 # v4.1.1 - 2021-08-27