+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-# Unreleased
+# v5.0.0 - 2021-11-24
 
 Breaking
   * AZ-515: Option to use Azure CAF naming provider to name resources